backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.6
psycopg2-binary==2.9.6
